subvortex/miner/metagraph/requirements.txt
bittensor==9.6.0
bittensor-wallet==3.0.10
loguru==0.7.0
numpy==2.0.1
pydantic==2.8.2
python-dotenv==1.0.1
setuptools==70.0.0
redis==5.0.1
tomli==2.2.1
tabulate==0.9.0